refactor: migrate JDBC imports to exposed.v1 and clean up unused imports
Standardized JDBC imports to `exposed.v1` package and removed unused imports for improved clarity and consistency.
